Ensure that DataPipeDrainer doesn't call OnDataComplete in ctor.
Currently DataPipeDrainer starts reading data directly in its
constructor, which might complete synchronously and call the completion
callback before returning from the constructor.

This is problematic for clients that want to delete the drainer upon
completion, as we end up calling the destructor from the constructor,
which seems to explode in a crash.

The patch fixes this by separating a ::Start() method and only starting
reading data when it's called.

Change-Id: I2a56e04f0d134578442345c3429e829e281d5a75