Update use of mx_handle_wait_many().
This change is somewhat involved and affects the API of MessageLoop
because the kernel no longer produces an error for "unsatisfiable"
waits.  Moreover, it also no longer reports a result index.

Instead, clients are expected to handle signals such as PEER_CLOSED
explicitly.  So we now pass that burden on to implementors of
MessageLoopHandler.

To maintain compatibility with mojo::AsyncWaiter, we emulate Mojo's
precondition failed behavior in DefaultAsyncWaiter.  This will
be deleted as soon as FIDL is ready.

Change-Id: I807ec7b980fad86eab40c7a582c6f7cd740a1387