Fix handler starvation issue in MessageLoop.
This change causes incoming tasks to be reloaded only in response
to a port packet which ensures that all prior packets on the port
have been handled before accepting additional tasks into the queue.

Previously at most one packet would be handled for each cluster
of incoming tasks meaning that it could take a very long time to
handle all packets if the incoming task queue was kept busy.
The only way to catch up was to stop posting tasks (or crash).

Change-Id: Ic65f757af311d9819cbbc26eba2982372f4cf594