Change MessageLoop destructor ordering.
The previous implementation of MessageLoop first clears it delegate
before calling shutdown on the loop. Calling shutdown on the loop
triggers callbacks on all currently registered handler. If any of these
callbacks tries to check that they are running on the correct thread,
they will crash.

This CL changes the order of event so that the delegate is only removed
after all the handlers have been called.

Also added regression tests for this condition.

US-325 #done

Change-Id: I71fb17c946d95181c6bcc26c0a2d7b20d29841a9